requirements.txt
yfinance
python-telegram-bot
schedule
anthropic
python-dotenv
pandas
termcolor
